Update models to the latest versions

+++ src/model/CancellationPolicy.php
@@ -11,8 +11,8 @@ namespace hotelbeds\hotel_api_sdk\model;
 /**
  * Class CancellationPolicy
  * @package hotelbeds\hotel_api_sdk\model
- * @property float amount Amount that will be charged after date from
- * @property float hotelAmount Amount that will be charged after date from in hotel currency (for pay at hotel model)
+ * @property double amount Amount that will be charged after date from
+ * @property double hotelAmount Amount that will be charged after date from in hotel currency (for pay at hotel model)
  * @property string hotelCurrency Currency of amount
  * @property string from Date from where the amount will be charged
  */
@@ -22,8 +22,8 @@ class CancellationPolicy extends ApiModel
     public function __construct(array $data=null)
     {
         $this->validFields = [
-            "amount" => "float",
-            "hotelAmount" => "float",
+            "amount" => "double",
+            "hotelAmount" => "double",
             "hotelCurrency" => "string",
             "from" => "string"
         ];
